trainer/requirements.txt
torch>=2.3
torchvision>=0.18
timm>=1.0
numpy>=1.26
opencv-python-headless>=4.10
scikit-learn>=1.5
joblib>=1.4
tqdm>=4.66
pandas>=2.2
faiss-cpu>=1.8
